TechTree Overhaul Made a CTT replacement for when CTT is not present.
Changed parts to interact with CTT better for upgradetech etc.

Made some exceptions in the script for parts that would not be compatible to CTT otherwise.

Changed parts tech requirement to that of the official techmanager tech tree.

Created Dummy entries like mechjeb uses to fill the techtree with little bits of info when upgrades unlock.